Call our helper to set that.
Fun side note; the boolean was defaulting to false and never getting
updated that very first time when it should've swapped to true.

Not that anything actually USES _renderContinously, and it's private so
it's hard for anything to, so it should probably just go away, but I'll
ignore that for now.